refactor implant subpackage
