+++ pulse2percept/implants/tests/test_electrodes.py
@@ -1,0 +1,76 @@
+import numpy as np
+import pytest
+import numpy.testing as npt
+
+from pulse2percept.implants import Electrode, DiskElectrode, PointSource
+
+
+class ValidElectrode(Electrode):
+    __slots__ = ()
+
+    def electric_potential(self, x, y, z):
+        r = np.sqrt((x - self.x) ** 2 + (y - self.y) ** 2 + (z - self.z) ** 2)
+        return r
+
+
+def test_Electrode():
+    electrode = ValidElectrode(0, 1, 2)
+    npt.assert_almost_equal(electrode.x, 0)
+    npt.assert_almost_equal(electrode.y, 1)
+    npt.assert_almost_equal(electrode.z, 2)
+    npt.assert_almost_equal(electrode.electric_potential(0, 1, 2), 0)
+    with pytest.raises(TypeError):
+        ValidElectrode([0], 1, 2)
+    with pytest.raises(TypeError):
+        ValidElectrode(0, np.array([1, 2]), 2)
+    with pytest.raises(TypeError):
+        ValidElectrode(0, 1, [2, 3])
+    # Slots:
+    npt.assert_equal(hasattr(electrode, '__slots__'), True)
+    npt.assert_equal(hasattr(electrode, '__dict__'), False)
+
+
+def test_PointSource():
+    electrode = PointSource(0, 1, 2)
+    npt.assert_almost_equal(electrode.x, 0)
+    npt.assert_almost_equal(electrode.y, 1)
+    npt.assert_almost_equal(electrode.z, 2)
+    npt.assert_almost_equal(electrode.electric_potential(0, 1, 2, 1, 1), 1)
+    npt.assert_almost_equal(electrode.electric_potential(0, 0, 0, 1, 1), 0.035,
+                            decimal=3)
+    # Slots:
+    npt.assert_equal(hasattr(electrode, '__slots__'), True)
+    npt.assert_equal(hasattr(electrode, '__dict__'), False)
+
+
+def test_DiskElectrode():
+    with pytest.raises(TypeError):
+        DiskElectrode(0, 0, 0, [1, 2])
+    with pytest.raises(TypeError):
+        DiskElectrode(0, np.array([0, 1]), 0, 1)
+    # Invalid radius:
+    with pytest.raises(ValueError):
+        DiskElectrode(0, 0, 0, -5)
+    # Check params:
+    electrode = DiskElectrode(0, 1, 2, 100)
+    npt.assert_almost_equal(electrode.x, 0)
+    npt.assert_almost_equal(electrode.y, 1)
+    npt.assert_almost_equal(electrode.z, 2)
+    # On the electrode surface (z=2, x^2+y^2<=100^2)
+    npt.assert_almost_equal(electrode.electric_potential(0, 1, 2, 1), 1)
+    npt.assert_almost_equal(electrode.electric_potential(30, -30, 2, 1), 1)
+    npt.assert_almost_equal(electrode.electric_potential(0, 101, 2, 1), 1)
+    npt.assert_almost_equal(electrode.electric_potential(0, -99, 2, 1), 1)
+    npt.assert_almost_equal(electrode.electric_potential(100, 1, 2, 1), 1)
+    npt.assert_almost_equal(electrode.electric_potential(-100, 1, 2, 1), 1)
+    # Right off the surface (z=2, x^2+y^2>100^2)
+    npt.assert_almost_equal(electrode.electric_potential(0, 102, 2, 1), 0.910,
+                            decimal=3)
+    npt.assert_almost_equal(electrode.electric_potential(0, -100, 2, 1), 0.910,
+                            decimal=3)
+    # Some distance away from the electrode (z>2):
+    npt.assert_almost_equal(electrode.electric_potential(0, 1, 38, 1), 0.780,
+                            decimal=3)
+    # Slots:
+    npt.assert_equal(hasattr(electrode, '__slots__'), True)
+    npt.assert_equal(hasattr(electrode, '__dict__'), False)
